requirements.txt
gradio
pandas
numpy
Pillow
scipy
sentence-transformers
huggingface-hub
spacy
ultralytics
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
opencv-python-headless